+++ src/lang/nodes/Conditional.js
@@ -1,17 +1,39 @@
 export function compile(node, ctx) {
     ctx.put(ctx.buildinFn('bool'));
     ctx.put('(');
-    ctx.scope.captureCurrent.disabled = true;
     ctx.nodeOrCurrent(node.test);
-    ctx.put(')?');
-    ctx.nodeOrCurrent(node.consequent);
-    ctx.put(':');
-    if (node.alternate) {
-        ctx.node(node.alternate);
-    } else {
-        ctx.put('undefined');
-    }
-    ctx.scope.captureCurrent.disabled = false;
+    ctx.createScope(
+        () => {
+            ctx.put(')?');
+            ctx.nodeOrCurrent(node.consequent);
+        },
+        (scopeStart, sp) => {
+            ctx.put(')');
+            return scopeStart + '(' + sp + ',';
+        },
+        ctx.scope.$ref
+    );
+    ctx.createScope(
+        () => {
+            ctx.put(':');
+            if (node.alternate) {
+                if (node.alternate.type === 'Placeholder') {
+                    ctx.put('(');
+                    ctx.node(node.alternate);
+                    ctx.put(',undefined)');
+                } else {
+                    ctx.node(node.alternate);
+                }
+            } else {
+                ctx.put('undefined');
+            }
+        },
+        (scopeStart, sp) => {
+            ctx.put(')');
+            return scopeStart + '(' + sp + ',';
+        },
+        ctx.scope.$ref
+    );
 }
 export function walk(node, ctx) {
     ctx.nodeOrNothing(node.test);